Convert magns into floats

+++ cchloader/models/corbagen.py
@@ -9,8 +9,8 @@ class CorbaGenSchema(Schema):
     name = fields.String(position=0, required=True)
     local_timestamp = fields.DateTime(position=1, format='%Y-%m-%d %H:%M', required=True)
     season = fields.Integer(position=2, validate=OneOf([0, 1]), required=True)
-    generacio = fields.Integer(position=3, required=True)
-    autocons = fields.Integer(position=4, required=True)
-    excedent = fields.Integer(position=5, required=True)
+    generacio = fields.Float(position=3, required=True)
+    autocons = fields.Float(position=4, required=True)
+    excedent = fields.Float(position=5, required=True)
 
 CorbaGenSchema()
